requirements.txt
dify_plugin==0.5.0
watercrawl-py==0.9.2